# Optimize CronExpression and small reduction on allocation