Micro performance optimization in CronExpression

**4.x with this change:**

|            Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|----------:|
| CreateExpressions | 35.96 us | 0.505 us | 0.518 us | 3.4180 |  55.93 KB |

4.x before change:  0c9ba23189c85223917d1ecbebbf2a951b891044

|            Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|----------:|
| CreateExpressions | 50.90 us | 0.981 us | 1.008 us | 5.0049 |  82.08 KB |

3.x 9d136a8d26122f0ff087f7ff67705673c1742ce8

|            Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|------------------ |---------:|---------:|---------:|-------:|----------:|
| CreateExpressions | 41.09 us | 0.783 us | 0.837 us | 4.5776 |  75.13 KB |


~~Branched from #1980, will rebase if/when PR 1980 is committed.~~

EDIT:

Benchmark now parameterised, example result below:

|                                     Method |                Date |       CronExpression |     Mean |    Error |   StdDev |   Gen0 |   Gen1 | Allocated |
|------------------------------------------- |-------------------- |--------------------- |---------:|---------:|---------:|-------:|-------:|----------:|
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |  0 0/5 10 6,15 * ? * | 13.12 us | 0.217 us | 0.203 us | 0.8392 | 0.0153 |  13.88 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 (...)* ? * [39] | 12.27 us | 0.118 us | 0.105 us | 0.8087 | 0.0153 |  13.29 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 (...)* ? * [24] | 13.68 us | 0.273 us | 0.464 us | 0.9003 | 0.0153 |  14.84 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |     0 15 10 15 * ? * | 13.06 us | 0.249 us | 0.306 us | 0.8545 | 0.0153 |  14.15 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |  0 15 10 15,31 * ? * | 13.16 us | 0.259 us | 0.425 us | 0.8850 |      - |  14.53 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |  0 15 10 15,31 * ? * | 13.35 us | 0.265 us | 0.284 us | 0.8850 | 0.0153 |  14.53 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |   0 15 10 15,L * ? * | 13.07 us | 0.212 us | 0.188 us | 0.8545 | 0.0153 |  14.19 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 10 15,L-2 * ? * | 13.08 us | 0.258 us | 0.297 us | 0.8698 | 0.0153 |  14.31 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 (...)* ? * [21] | 13.74 us | 0.272 us | 0.373 us | 0.8850 | 0.0153 |   14.6 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |   0 15 10 6,15 * ? * | 12.89 us | 0.202 us | 0.189 us | 0.8698 | 0.0153 |  14.41 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 10 6,15,L * ? * | 12.89 us | 0.197 us | 0.184 us | 0.8698 | 0.0153 |  14.43 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 | 0 15 (...)* ? * [21] | 12.95 us | 0.236 us | 0.242 us | 0.8698 | 0.0153 |  14.44 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 06/01/2005 22:15:00 |    0 15 10 ? * 6#3 * | 12.57 us | 0.145 us | 0.129 us | 0.7935 | 0.0153 |  13.13 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |  0 0/5 10 6,15 * ? * | 13.03 us | 0.256 us | 0.342 us | 0.8240 | 0.0153 |  13.62 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 (...)* ? * [39] | 12.74 us | 0.250 us | 0.411 us | 0.8240 | 0.0153 |  13.55 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 (...)* ? * [24] | 11.96 us | 0.238 us | 0.284 us | 0.7477 | 0.0153 |   12.4 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |     0 15 10 15 * ? * | 13.02 us | 0.169 us | 0.158 us | 0.8698 | 0.0153 |  14.41 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |  0 15 10 15,31 * ? * | 13.35 us | 0.238 us | 0.223 us | 0.9003 | 0.0153 |   14.8 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |  0 15 10 15,31 * ? * | 13.07 us | 0.218 us | 0.204 us | 0.9003 | 0.0153 |   14.8 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |   0 15 10 15,L * ? * | 12.90 us | 0.172 us | 0.161 us | 0.8698 | 0.0153 |  14.45 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 10 15,L-2 * ? * | 13.16 us | 0.158 us | 0.140 us | 0.8850 | 0.0153 |  14.58 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 (...)* ? * [21] | 13.50 us | 0.254 us | 0.260 us | 0.9003 | 0.0153 |  14.87 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |   0 15 10 6,15 * ? * | 13.04 us | 0.200 us | 0.187 us | 0.8850 | 0.0153 |  14.68 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 10 6,15,L * ? * | 13.09 us | 0.106 us | 0.099 us | 0.8850 | 0.0153 |   14.7 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 | 0 15 (...)* ? * [21] | 13.49 us | 0.258 us | 0.286 us | 0.8850 | 0.0153 |   14.7 KB |
| CreateExpressionsAndCalculateFireTimeAfter | 12/31/2005 23:59:59 |    0 15 10 ? * 6#3 * | 12.54 us | 0.109 us | 0.102 us | 0.8087 | 0.0153 |  13.39 KB |